bot/requirements.txt
python-telegram-bot==22.0
python-telegram-bot[job-queue]
pytz==2025.2
